fix(schemas): replace float with Decimal for money fields across finance, cafeteria, hr, inventory, library
- finance.py: converted 28 money fields (debit, credit, amount, balance, total_amount, paid_amount, salary, allowances, etc.)
- cafeteria.py: converted price and total fields
- hr.py: converted basic_salary field
- inventory.py: converted quantity, min_quantity, unit_price fields
- library.py: converted fine_amount field
- TECHNICAL_DEBT.md: marked float-money item RESOLVED
- CHANGELOG.md: updated with 0.9.2 release notes

Refs: Deep Audit 2026-07-06 float money contract violation

+++ backend/app/schemas/cafeteria.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from pydantic import BaseModel, Field
 from typing import Optional
 from datetime import datetime
@@ -5,11 +6,11 @@ from datetime import datetime
 
 class ProductCreate(BaseModel):
     name: str = Field(max_length=255)
-    price: float; category: Optional[str] = None; stock: int = 0
+    price: Decimal; category: Optional[str] = None; stock: int = 0
 
 
 class ProductResponse(BaseModel):
-    id: str; name: str; price: float; category: Optional[str] = None
+    id: str; name: str; price: Decimal; category: Optional[str] = None
     stock: int; school_id: str; created_at: Optional[datetime] = None
 
 
@@ -19,7 +20,7 @@ class OrderItemCreate(BaseModel):
 
 class ProductUpdate(BaseModel):
     name: Optional[str] = None
-    price: Optional[float] = None
+    price: Optional[Decimal] = None
     category: Optional[str] = None
     stock: Optional[int] = None
 
@@ -37,5 +38,5 @@ class OrderCreate(BaseModel):
 
 class OrderResponse(BaseModel):
     id: str; customer_type: str; customer_id: Optional[str] = None
-    total: float; status: str; payment_method: Optional[str] = None
+    total: Decimal; status: str; payment_method: Optional[str] = None
     school_id: str; created_by: str; created_at: Optional[datetime] = None
